kernel: improve UART and logging.
Add different working modes for the UART. It allows to use the UART in
polled (no interrupt) or buffered (buffers and interrupts) modes to be
able to start logging before the interrupts are enabled.

The kernel is now able to start printing traces before the interrupts
are initialized. This CL also add a bunch of traces during the
initializtion of memory, uart and clock modules.