kernel: let console handle the buffering.
To allow the console to receive traces even before the UART (or any
other output) is initialized, we need to move all the buffering logic
in the console implementation itself. The console now handles the RX and
TX buffers and just interacts with the driver to provide the queues and
kick off transmission.