kernel: fix console_getchar() to return a byte.
console_getchar() is expected to return a character received on the
console. Since a char is not enough to return a character or an error
value, console_getchar() returns an int.

This CL fixes the way a real character is returned to avoid sign
extension from char to int, and avoid the confusion between a high char
value and an error value.